requirements.txt
Django==4.2.7                           #Основной фреймворк
djangorestframework==3.14.0             #Для работы с REST API
djangorestframework-simplejwt==5.3.0    #Для работы с JWT-токенами
psycopg2-binary==2.9.9                  #Для работы с PostgreSQL
drf-spectacular==0.27.                  #Для автоматической документации
python-decouple==3.8.0                  #Для хранения конфиденциальной информации (ключи, пароли и т.д.)
Pillow==10.1.0                          #Для работы с изображениями
django-cors-headers==4.3.1              #Для работы с CORS
djangorestframework-simplejwt           #Для импортирования модуля Simple JWT